+++ Dockerfile
@@ -1,0 +1,17 @@
+FROM python:3.9-slim
+
+WORKDIR /app
+
+COPY requirements.txt .
+RUN pip install --no-cache-dir -r requirements.txt
+
+COPY . .
+
+# Create templates directory if it doesn't exist
+RUN mkdir -p templates
+
+# Make port 8080 available
+EXPOSE 8080
+
+# Run the application with gunicorn
+CMD ["gunicorn", "--bind", "0.0.0.0:8080", "app:app"]
