+++ README.md
@@ -1,51 +1,59 @@
-# Fantasy OWCS Score Calc
+# Fantasy OWCS Score Calculator - Web Application
 
-Generate a match report from a link to a FACEIT Overwatch match that also contains scores for the [Fantasy OWCS web app](https://owfantasy.com/).
+This web application allows users to easily generate Fantasy OWCS score reports by simply pasting a FACEIT match URL. The application processes the match data and displays a formatted report directly in the browser, which can be easily copied to share on Discord or other platforms.
+
+## Features
+
+- Simple web interface for inputting FACEIT match URLs
+- Automatic processing of match data using the FACEIT API
+- Formatted display of match results and player scores
+- Copy to clipboard functionality for easy sharing
+- No installation required for end users
 
 ## Installation
 
 1. **Clone the repository:**
 
     ```sh
-    git clone git@github.com:UmarUMahmood/fantasy-owcs-score-calc.git
-    cd fantasy-owcs-score-calc
+    git clone git@github.com:YourUsername/fantasy-owcs-webapp.git
+    cd fantasy-owcs-webapp
     ```
 
 2. **Create a virtual environment:**
 
     ```sh
     python3 -m venv .venv
-    source .venv/bin/activate
+    source .venv/bin/activate  # On Windows: .venv\Scripts\activate
     ```
 
 3. **Install the required packages:**
 
     ```sh
-    pip install -r requirements.txt
+    pip install flask dotenv requests table2ascii
     ```
 
-## Configuration
-
-1. **Create a `.env` file with your FACEIT API key:**
+4. **Create a `.env` file with your FACEIT API key:**
 
     ```sh
-    echo "API_KEY=api-key-here" > .env
+    echo "API_KEY=your-api-key-here" > .env
     ```
 
-    Replace `api-key-here` with your actual FACEIT API key. You can obtain an API key from the [FACEIT Developer Portal](https://developers.faceit.com/).
+    Replace `your-api-key-here` with your actual FACEIT API key from the [FACEIT Developer Portal](https://developers.faceit.com/).
 
-## Usage
-
-1. **Run the script:**
+5. **Create the templates directory and add the HTML template:**
 
     ```sh
-    python3 main.py
+    mkdir -p templates
+    # Copy the index.html file into this directory
     ```
 
-2. **Copy/Paste the link to the match from FACEIT:**
+## Running the Application Locally
 
-    For example: `https://www.faceit.com/en/ow2/room/1-d260e7c6-0235-417e-80bf-40b41a0e6f60`
+1. **Start the Flask app:**
 
-3. **Copy/Paste the output:**
+    ```sh
+    python app.py
+    ```
 
-    After running the script, copy the output from the generated .md file in the `output` directory and paste it into Discord or wherever.
+2. **Access the application:**
+   Open your browser and navigate to `http://127.0.0.1:5000`
