+++ requirements.txt
@@ -1,4 +1,5 @@
+flask
 python-dotenv
 requests
-flask
 table2ascii
+gunicorn
