+++ templates/index.html
@@ -1,0 +1,228 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Fantasy OWCS Score Calculator</title>
+    <style>
+        html {
+            color-scheme: dark;
+        }
+
+        body {
+            font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+            line-height: 1.6;
+            color: #ffffff;
+            max-width: 1200px;
+            margin: 0 auto;
+            padding: 20px;
+            background-color: #181a1b;
+        }
+        
+        h1 {
+            color: #ec8302;
+            text-align: center;
+            margin-bottom: 30px;
+        }
+        
+        .container {
+            background-color: #131516;
+            padding: 30px;
+            border-radius: 8px;
+            box-shadow: 0 2px 10px rgba(219, 218, 218, 0.1);
+        }
+        
+        form {
+            margin-bottom: 30px;
+        }
+        
+        label {
+            display: block;
+            margin-bottom: 10px;
+            font-weight: bold;
+        }
+        
+        input[type="text"] {
+            width: 100%;
+            padding: 12px;
+            margin-bottom: 20px;
+            border: 1px solid #3a3e41;
+            border-radius: 4px;
+            box-sizing: border-box;
+            font-size: 16px;
+        }
+        
+        button {
+            background-color: #A35A00;
+            color: white;
+            border: none;
+            padding: 12px 20px;
+            cursor: pointer;
+            border-radius: 4px;
+            font-size: 16px;
+            transition: background-color 0.3s;
+        }
+        
+        button:hover {
+            background-color: rgb(243, 135, 2);
+        }
+        
+        #loading {
+            display: none;
+            text-align: center;
+            margin: 20px 0;
+        }
+        
+        #result {
+            background-color: #181818;
+            padding: 20px;
+            border-radius: 4px;
+            white-space: pre;
+            font-family: monospace;
+            overflow-x: auto;
+            display: block;
+            border: 1px solid #ddd;
+            margin-bottom: 15px;
+            text-align: center;
+        }
+        
+        .copy-btn {
+            margin-top: 15px;
+            background-color: #6c757d;
+        }
+        
+        .copy-btn:hover {
+            background-color: #5a6268;
+        }
+        
+        .success-message {
+            color: green;
+            margin-left: 10px;
+            display: none;
+        }
+        
+        .option-container {
+            margin-bottom: 20px;
+        }
+        
+        .checkbox-label {
+            display: inline;
+            font-weight: normal;
+            margin-left: 8px;
+        }
+        
+        .side-by-side-container {
+            display: flex;
+            flex-wrap: wrap;
+            gap: 20px;
+            margin-bottom: 20px;
+        }
+        
+        footer {
+            margin-top: 40px;
+            text-align: center;
+            color: #777;
+            font-size: 14px;
+        }
+
+        a { 
+            color: #3391ff;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h1>Fantasy OWCS Score Calculator</h1>
+
+        <div class="match-links" style="margin-bottom: 20px; text-align: center;">
+            Get a match link from FACEIT here:
+            <a href="https://www.faceit.com/en/ow2/league/FACEIT%20League/88c7f7ec-4cb8-44d3-a5db-6e808639c232/19cfb5d9-6a70-45cc-904f-2e8ef6f80a4b/matches?region=65b085e4-7e28-49f3-a6c1-9e12faaa343b&division=47812350-e0c6-48fd-8ff9-19d184236957&stage=ccea29fe-582e-4619-8afa-858bacc89a4f&conference=6306d719-8c2d-471c-bc42-421ada64911b" target="_blank">FACEIT EMEA Matches</a> | 
+            <a href="https://www.faceit.com/en/ow2/league/FACEIT%20League/88c7f7ec-4cb8-44d3-a5db-6e808639c232/19cfb5d9-6a70-45cc-904f-2e8ef6f80a4b/matches?region=e80b15e5-92eb-4500-8814-2353162a4727&division=d9e85be4-5622-430f-9f3f-f2f39485d28b&stage=ad5d41df-d838-4358-9cb0-dc464b79624e&conference=83cf0d5b-329b-4ac7-80c0-282238c336e3" target="_blank">FACEIT NA Matches</a>
+        </div>
+
+        <form id="matchForm">
+            <label for="matchUrl">Paste FACEIT Match URL:</label>
+            <input type="text" id="matchUrl" name="match_url" placeholder="https://www.faceit.com/en/ow2/room/1-xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx" required>
+            
+            <div class="option-container">
+                <input type="checkbox" id="sideBySide" name="side_by_side">
+                <label for="sideBySide" class="checkbox-label">Display team tables side by side</label>
+            </div>
+            
+            <button type="submit">Generate Report</button>
+        </form>
+        
+        <div id="loading">
+            <p>Processing match data... This may take a few moments.</p>
+        </div>
+        
+        <div id="result-container" style="display: none;">
+            <div id="result"></div>
+            <div style="text-align: right;">
+                <button id="copyBtn" class="copy-btn">Copy to Clipboard</button>
+                <span id="copySuccess" class="success-message">Copied!</span>
+            </div>
+        </div>
+    </div>
+    
+    <footer>
+        <p>Fantasy OWCS Score Calculator &copy; 2025 | <a href="https://owfantasy.com/" target="_blank">Visit Fantasy OWCS</a></p>
+    </footer>
+    
+    <script>
+        document.getElementById('matchForm').addEventListener('submit', function(e) {
+            e.preventDefault();
+            
+            const matchUrl = document.getElementById('matchUrl').value;
+            const sideBySide = document.getElementById('sideBySide').checked;
+            const loadingDiv = document.getElementById('loading');
+            const resultContainer = document.getElementById('result-container');
+            const resultDiv = document.getElementById('result');
+            
+            // Show loading message
+            loadingDiv.style.display = 'block';
+            resultContainer.style.display = 'none';
+            resultDiv.textContent = '';
+            
+            // Send request to server
+            fetch('/process', {
+                method: 'POST',
+                headers: {
+                    'Content-Type': 'application/x-www-form-urlencoded',
+                },
+                body: new URLSearchParams({
+                    'match_url': matchUrl,
+                    'side_by_side': sideBySide
+                })
+            })
+            .then(response => response.json())
+            .then(data => {
+                loadingDiv.style.display = 'none';
+                resultContainer.style.display = 'block';
+                
+                if (data.error) {
+                    resultDiv.textContent = 'Error: ' + data.error;
+                } else {
+                    resultDiv.textContent = data.report;
+                }
+            })
+            .catch(error => {
+                loadingDiv.style.display = 'none';
+                resultContainer.style.display = 'block';
+                resultDiv.textContent = 'Error processing request: ' + error;
+            });
+        });
+        
+        document.getElementById('copyBtn').addEventListener('click', function() {
+            const resultText = document.getElementById('result').textContent;
+            navigator.clipboard.writeText(resultText).then(function() {
+                const successMsg = document.getElementById('copySuccess');
+                successMsg.style.display = 'inline';
+                setTimeout(function() {
+                    successMsg.style.display = 'none';
+                }, 2000);
+            });
+        });
+    </script>
+</body>
+</html>
